fix int64 error on json output

+++ lddecode_core.py
@@ -1018,11 +1018,11 @@ class Field:
 
                 start_rf_linepos = curerr[0] - self.linelocs[l]
                 start_linepos = start_rf_linepos / (self.linelocs[l + 1] - self.linelocs[l])
-                start_linepos = int(np.round(start_linepos * self.outlinelen))
+                start_linepos = np.round(start_linepos * self.outlinelen)
 
                 end_rf_linepos = curerr[1] - self.linelocs[l]
                 end_linepos = end_rf_linepos / (self.linelocs[l + 1] - self.linelocs[l])
-                end_linepos = int(np.round(end_linepos * self.outlinelen))
+                end_linepos = np.round(end_linepos * self.outlinelen)
                 
                 if end_linepos > self.outlinelen:
                     # need to output two dropouts
